SCRUM-151 refactor: 에디터, 파일 업로드, 첨부파일 뷰 공통 컴포넌트 화
